backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
curl-cffi==0.13.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dateutil==2.9.0.post0
cachetools==5.5.0
mercantile==1.2.1
